Move Main service to services/redux/index.ts
This is the first step to better encapsulating the responsibilities of
the main service as specifically a Redux service.

No changes are introduced to the service itself, it's just moved
alongside one constant that is promoted to background constants.
