+++ background/redux-slices/utils.ts
@@ -10,7 +10,7 @@ import logger from "../lib/logger"
 // FIXME: this utility file should not depend on actual services.
 // Creating a properly typed version of `createBackgroundAsyncThunk`
 // elsewhere has proven quite hard, so we are hardwiring typing here.
-import type Main from "../main"
+import type ReduxService from "../services/redux"
 
 // Below, we use `any` to deal with the fact that allAliases is a heterogeneous
 // collection of async thunk actions whose payload types have little in common
@@ -87,7 +87,7 @@ export type BackgroundAsyncThunk<
   TypePrefix extends string,
   Returned,
   ThunkArg = void,
-  ThunkApiConfig extends AsyncThunkConfig = { extra: { main: Main } },
+  ThunkApiConfig extends AsyncThunkConfig = { extra: { main: ReduxService } },
 > = ((
   payload: ThunkArg,
 ) => BackgroundAsyncThunkAction<TypePrefix, Returned> & { payload: ThunkArg }) &
@@ -143,7 +143,7 @@ export function createBackgroundAsyncThunk<
   TypePrefix extends string,
   Returned,
   ThunkArg = void,
-  ThunkApiConfig extends AsyncThunkConfig = { extra: { main: Main } },
+  ThunkApiConfig extends AsyncThunkConfig = { extra: { main: ReduxService } },
 >(
   typePrefix: TypePrefix,
   payloadCreator: AsyncThunkPayloadCreator<Returned, ThunkArg, ThunkApiConfig>,
