+++ background/services/redux/index.ts
@@ -15,7 +15,7 @@ import {
   normalizeEVMAddress,
   sameEVMAddress,
   wait,
-} from "./lib/utils"
+} from "../../lib/utils"
 
 import {
   BaseService,
@@ -36,14 +36,14 @@ import {
   WalletConnectService,
   AnalyticsService,
   getNoopService,
-} from "./services"
+} from ".."
 
-import { HexString, NormalizedEVMAddress } from "./types"
-import { SignedTransaction } from "./networks"
-import { AccountBalance, AddressOnNetwork, NameOnNetwork } from "./accounts"
-import { Eligible, ReferrerStats } from "./services/island/types"
+import { HexString, NormalizedEVMAddress } from "../../types"
+import { SignedTransaction } from "../../networks"
+import { AccountBalance, AddressOnNetwork, NameOnNetwork } from "../../accounts"
+import { Eligible, ReferrerStats } from "../island/types"
 
-import rootReducer from "./redux-slices"
+import rootReducer from "../../redux-slices"
 import {
   AccountType,
   deleteAccount,
@@ -51,27 +51,27 @@ import {
   updateAccountBalance,
   updateAccountName,
   updateENSAvatar,
-} from "./redux-slices/accounts"
+} from "../../redux-slices/accounts"
 import {
   assetsLoaded,
   refreshAsset,
   removeAssetData,
-} from "./redux-slices/assets"
+} from "../../redux-slices/assets"
 import {
   addIslandAsset,
   setEligibility,
   setEligibilityLoading,
   setReferrer,
   setReferrerStats,
-} from "./redux-slices/claim"
+} from "../../redux-slices/claim"
 import {
   emitter as internalSignerSliceEmitter,
   internalSignerLocked,
   internalSignerUnlocked,
   updateInternalSigners,
   setKeyringToVerify,
-} from "./redux-slices/internal-signer"
-import { blockSeen, setEVMNetworks } from "./redux-slices/networks"
+} from "../../redux-slices/internal-signer"
+import { blockSeen, setEVMNetworks } from "../../redux-slices/networks"
 import {
   initializationLoadingTimeHitLimit,
   emitter as uiSliceEmitter,
@@ -87,7 +87,7 @@ import {
   toggleNotifications,
   setShownDismissableItems,
   dismissableItemMarkedAsShown,
-} from "./redux-slices/ui"
+} from "../../redux-slices/ui"
 import {
   estimatedFeesPerGas,
   emitter as transactionConstructionSliceEmitter,
@@ -99,16 +99,16 @@ import {
   transactionSigned,
   clearCustomGas,
   updateRollupEstimates,
-} from "./redux-slices/transaction-construction"
-import { selectDefaultNetworkFeeSettings } from "./redux-slices/selectors/transactionConstructionSelectors"
-import { allAliases } from "./redux-slices/utils"
+} from "../../redux-slices/transaction-construction"
+import { selectDefaultNetworkFeeSettings } from "../../redux-slices/selectors/transactionConstructionSelectors"
+import { allAliases } from "../../redux-slices/utils"
 import {
   requestPermission,
   emitter as providerBridgeSliceEmitter,
   initializePermissions,
   revokePermissionsForAddress,
-} from "./redux-slices/dapp"
-import logger from "./lib/logger"
+} from "../../redux-slices/dapp"
+import logger from "../../lib/logger"
 import {
   rejectDataSignature,
   clearSigningState,
@@ -117,52 +117,64 @@ import {
   signingSliceEmitter,
   typedDataRequest,
   signDataRequest,
-} from "./redux-slices/signing"
+} from "../../redux-slices/signing"
 
-import { SignTypedDataRequest, MessageSigningRequest } from "./utils/signing"
+import {
+  SignTypedDataRequest,
+  MessageSigningRequest,
+} from "../../utils/signing"
 import {
   emitter as earnSliceEmitter,
   setVaultsAsStale,
-} from "./redux-slices/earn"
+} from "../../redux-slices/earn"
 import {
   removeDevice,
   resetLedgerState,
   setDeviceConnectionStatus,
   setUsbDeviceCount,
-} from "./redux-slices/ledger"
-import { ETHEREUM, FLASHBOTS_RPC_URL, OPTIMISM, USD } from "./constants"
-import { clearApprovalInProgress, clearSwapQuote } from "./redux-slices/0x-swap"
+} from "../../redux-slices/ledger"
+import {
+  ETHEREUM,
+  FLASHBOTS_RPC_URL,
+  OPTIMISM,
+  USD,
+  POPUP_MONITOR_PORT_NAME,
+} from "../../constants"
+import {
+  clearApprovalInProgress,
+  clearSwapQuote,
+} from "../../redux-slices/0x-swap"
 import {
   AccountSigner,
   SignatureResponse,
   TXSignatureResponse,
-} from "./services/signing"
+} from "../signing"
 import {
   migrateReduxState,
   REDUX_STATE_VERSION,
-} from "./redux-slices/migrations"
-import { PermissionMap } from "./services/provider-bridge/utils"
-import { TAHO_INTERNAL_ORIGIN } from "./services/internal-ethereum-provider/constants"
+} from "../../redux-slices/migrations"
+import { PermissionMap } from "../provider-bridge/utils"
+import { TAHO_INTERNAL_ORIGIN } from "../internal-ethereum-provider/constants"
 import {
   ActivityDetail,
   addActivity,
   initializeActivities,
   initializeActivitiesForAccount,
   removeActivities,
-} from "./redux-slices/activities"
-import { selectActivitesHashesForEnrichment } from "./redux-slices/selectors"
-import { getActivityDetails } from "./redux-slices/utils/activities-utils"
-import { getRelevantTransactionAddresses } from "./services/enrichment/utils"
-import { AccountSignerWithId } from "./signing"
+} from "../../redux-slices/activities"
+import { selectActivitesHashesForEnrichment } from "../../redux-slices/selectors"
+import { getActivityDetails } from "../../redux-slices/utils/activities-utils"
+import { getRelevantTransactionAddresses } from "../enrichment/utils"
+import { AccountSignerWithId } from "../../signing"
 import {
   AnyAssetMetadata,
   assetAmountToDesiredDecimals,
   convertAssetAmountViaPricePoint,
   isSmartContractFungibleAsset,
   SmartContractFungibleAsset,
-} from "./assets"
-import { FeatureFlags, isEnabled } from "./features"
-import { NFTCollection } from "./nfts"
+} from "../../assets"
+import { FeatureFlags, isEnabled } from "../../features"
+import { NFTCollection } from "../../nfts"
 import {
   initializeNFTs,
   updateNFTsCollections,
@@ -171,8 +183,8 @@ import {
   deleteNFTsForAddress,
   updateIsReloading,
   deleteTransferredNFTs,
-} from "./redux-slices/nfts"
-import AbilitiesService from "./services/abilities"
+} from "../../redux-slices/nfts"
+import AbilitiesService from "../abilities"
 import {
   setAbilitiesForAddress,
   updateAbility,
@@ -180,26 +192,23 @@ import {
   deleteAccount as deleteAccountFilter,
   deleteAbilitiesForAccount,
   initAbilities,
-} from "./redux-slices/abilities"
-import { AddChainRequestData } from "./services/provider-bridge"
+} from "../../redux-slices/abilities"
+import { AddChainRequestData } from "../provider-bridge"
 import {
   AnalyticsEvent,
   isOneTimeAnalyticsEvent,
   OneTimeAnalyticsEvent,
-} from "./lib/posthog"
+} from "../../lib/posthog"
 import {
   isBaseAssetForNetwork,
   isSameAsset,
-} from "./redux-slices/utils/asset-utils"
-import {
-  SignerImportMetadata,
-  SignerInternalTypes,
-} from "./services/internal-signer"
-import { getPricePoint, getTokenPrices } from "./lib/prices"
-import { makeFlashbotsProviderCreator } from "./services/chain/serial-fallback-provider"
-import { AnalyticsPreferences, DismissableItem } from "./services/preferences"
-import { newPricePoints } from "./redux-slices/prices"
-import NotificationsService from "./services/notifications"
+} from "../../redux-slices/utils/asset-utils"
+import { SignerImportMetadata, SignerInternalTypes } from "../internal-signer"
+import { getPricePoint, getTokenPrices } from "../../lib/prices"
+import { makeFlashbotsProviderCreator } from "../chain/serial-fallback-provider"
+import { AnalyticsPreferences, DismissableItem } from "../preferences"
+import { newPricePoints } from "../../redux-slices/prices"
+import NotificationsService from "../notifications"
 
 // This sanitizer runs on store and action data before serializing for remote
 // redux devtools. The goal is to end up with an object that is directly
@@ -290,8 +299,6 @@ const initializeStore = (main: Main, preloadedState: object) =>
 
 type ReduxStoreType = ReturnType<typeof initializeStore>
 
-export const popupMonitorPortName = "popup-monitor"
-
 // TODO Rename ReduxService or CoordinationService, move to services/, etc.
 export default class Main extends BaseService<never> {
   /**
@@ -376,6 +383,7 @@ export default class Main extends BaseService<never> {
 
       if (state) {
         const restoredState = decodeJSON(state)
+
         if (typeof restoredState === "object" && restoredState !== null) {
           // If someone managed to sneak JSON that decodes to typeof "object"
           // but isn't a Record<string, unknown>, there is a very large
@@ -2059,7 +2067,7 @@ export default class Main extends BaseService<never> {
 
   private connectPopupMonitor() {
     runtime.onConnect.addListener((port) => {
-      if (port.name !== popupMonitorPortName) return
+      if (port.name !== POPUP_MONITOR_PORT_NAME) return
 
       const openTime = Date.now()
